pyproject.toml
[project]
name = "mole-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.1",
    "mcp>=1.11.0",
    "python-dotenv>=1.1.1",
    "pyjwt>=2.8.0",
    "pydantic>=2.5.0",
]

[project.scripts]
stealthmole-mcp = "stealthmole_mcp.server:main"
mole-mcp-server = "stealthmole_mcp.server:main"
